evals/sensor_validity.py
"""
Eval: Sensor Validity
---------------------
Verifies that the agent correctly identifies out-of-range or implausible
sensor readings in a snapshot and includes them in its assessment.

Scoring: binary per case (1 = agent flagged the bad sensor, 0 = missed it)
"""

import os
import json
import pytest
from openai import OpenAI
from dotenv import load_dotenv

load_dotenv()
client = OpenAI(api_key=os.environ["OPENAI_API_KEY"])

# Each case: snapshot with one intentionally bad sensor + expected flag keyword
SENSOR_VALIDITY_CASES = [
    {
        "id": "ecg_overcool",
        "vehicleId": "tundra-2007",
        "snapshot": {"RPM": 800, "ECT": 110, "STFT_B1": 0.0, "LTFT_B1": 0.0},
        "expect_flag": "ECT",
        "description": "Engine coolant temp too low (thermostat stuck open)",
    },
    {
        "id": "maf_zero",
        "vehicleId": "tundra-2007",
        "snapshot": {"RPM": 750, "ECT": 195, "MAF": 0.0, "STFT_B1": 0.0},
        "expect_flag": "MAF",
        "description": "MAF reads zero at idle — sensor failure or unplugged",
    },
    {
        "id": "ltft_extreme_lean",
        "vehicleId": "bmw-335i-2009",
        "snapshot": {"RPM": 800, "ECT": 200, "LTFT_B1": 24.5, "STFT_B1": 9.0},
        "expect_flag": "LTFT",
        "description": "LTFT B1 near saturation — severe lean condition",
    },
    {
        "id": "rpm_unrealistic",
        "vehicleId": "honda-fit-2015",
        "snapshot": {"RPM": 9999, "ECT": 190, "STFT_B1": 0.0, "LTFT_B1": 0.0},
        "expect_flag": "RPM",
        "description": "RPM above redline — likely sensor error",
    },
]

SYSTEM_PROMPT = (
    "You are an OBD2 diagnostic assistant. "
    "Analyze the snapshot and list any sensors with implausible or out-of-range values."
)


def call_agent(snapshot: dict) -> str:
    response = client.chat.completions.create(
        model="gpt-4o",
        messages=[
            {"role": "system", "content": SYSTEM_PROMPT},
            {
                "role": "user",
                "content": f"Snapshot:\n```json\n{json.dumps(snapshot, indent=2)}\n```",
            },
        ],
        temperature=0,
    )
    return response.choices[0].message.content or ""


@pytest.mark.parametrize("case", SENSOR_VALIDITY_CASES, ids=[c["id"] for c in SENSOR_VALIDITY_CASES])
def test_sensor_flagged(case):
    """Agent must mention the expected sensor name in its response."""
    response = call_agent(case["snapshot"])
    assert case["expect_flag"] in response, (
        f"Case '{case['id']}': expected '{case['expect_flag']}' to be flagged.\n"
        f"Agent response:\n{response}"
    )
